backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-Bcrypt==1.0.1
PyJWT==2.1.0
Werkzeug==3.1.1
flask-jwt-extended==4.7.1
flask-caching==2.3.1

